requirements.txt
resemblyzer==0.1.1.dev0
tensorboardX==2.5.1
SoundFile==0.11.0
h5py==3.7.0
future==0.18.2
g2p-en==2.1.0
g2pM==0.1.2.5
et-xmlfile==1.0.1
absl-py==1.3.0

altgraph==0.17.3
appdirs==1.4.4
async-timeout==4.0.2
audioread==3.0.0
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
scipy==1.9.3
numpy==1.23.5
librosa==0.9.1
matplotlib==3.6.2
torchcrepe==0.0.17
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-lightning==0.7.1
six==1.16.0
tqdm==4.64.1
resampy==0.4.2
imageio==2.23.0
einops==0.6.0
pycwt==0.3.0a22
praat-parselmouth==0.4.3
scikit-image==0.19.3
pyloudnorm==0.1.0
torchmetrics==0.5.0
tensorboard==2.11.0
tensorboard-plugin-wit==1.8.1
protobuf==3.13.0
PyYAML==5.4
pypinyin==0.39.0

# It is recommended to install PyTorch manually.
# See instructions at https://pytorch.org/get-started/previous-versions/
# torch==1.8.2
# torchaudio==0.8.2
# torchvision==0.9.2
